+++ src/NFTCard.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { transferNft } from './utils/interact';
+import { transferNft } from './utils/productInteractions';
 
 interface NFTCardProps {
   nft: any
@@ -9,7 +9,7 @@ const NFTCard:React.FC<NFTCardProps> = ({ nft }) => {
 
   return (
     <div style={{ backgroundColor: "white", borderRadius: "20px", width: "50%", padding: "10px", margin: "0px auto 15px auto" }}>
-      <img alt="nft-img" src={nft.image_url} />
+      <img alt="nft-img" src={nft.image_url} style={{ maxWidth: "80%" }}/>
       <p>Collection: {nft.collection.slug}</p>
       <p>Contract: {nft.asset_contract.address}, tokenId: {nft.token_id}</p>
       <p>Name: {nft.name}</p>
